Added all methods from BOARD to CHUNKBOARD; removed old BOARD and renamed CHUNKBOARD to BOARD!!; Fixes #5
